test runtime info fixtures extracted to separate module
