+++ tests/test_app_platform.py
@@ -7,7 +7,7 @@ from pykube import ConfigMap
 from pytest_mock import MockerFixture
 
 import pytest_helm_charts
-import pytest_helm_charts.fixtures
+import pytest_helm_charts.test_runtime_fixtures
 import pytest_helm_charts.giantswarm_app_platform.utils
 import pytest_helm_charts.giantswarm_app_platform.fixtures
 from pytest_helm_charts.clusters import Cluster
